+++ lib/pages/search_page.dart
@@ -1,0 +1,77 @@
+import 'package:converterpro/utils/property_unit_list.dart';
+import 'package:converterpro/utils/utils_widgets.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+
+class CustomSearchDelegate extends SearchDelegate<int> {
+  final List<int> orderList;
+
+  CustomSearchDelegate(this.orderList);
+
+  @override
+  ThemeData appBarTheme(BuildContext context) => Theme.of(context);
+
+  @override
+  Widget buildLeading(BuildContext context) {
+    return IconButton(
+      tooltip: AppLocalizations.of(context)!.back,
+      icon: AnimatedIcon(
+        icon: AnimatedIcons.menu_arrow,
+        progress: transitionAnimation,
+      ),
+      onPressed: () {
+        close(context, 0);
+      },
+    );
+  }
+
+  @override
+  Widget buildSuggestions(BuildContext context) {
+    final Brightness brightness = Theme.of(context).brightness;
+
+    final List<SearchUnit> _dataSearch = getSearchUnitsList((int pageNumber) {
+      close(context, pageNumber);
+    }, context);
+    final List<SearchGridTile> allConversions = initializeGridSearch(
+      (int pageNumber) {
+        close(context, pageNumber);
+      },
+      context,
+      brightness == Brightness.dark,
+      orderList,
+    );
+
+    final Iterable<SearchUnit> suggestions = _dataSearch.where((searchUnit) =>
+        searchUnit.unitName.toLowerCase().contains(query.toLowerCase())); //.toLowercase in order to be case insesitive
+
+    return query.isNotEmpty
+        ? SuggestionList(
+            suggestions: suggestions.toList(),
+            darkMode: brightness == Brightness.dark,
+          )
+        : GridView(
+            gridDelegate: const SliverGridDelegateWithMaxCrossAxisExtent(maxCrossAxisExtent: 180.0),
+            children: allConversions,
+          );
+  }
+
+  @override
+  Widget buildResults(BuildContext context) {
+    return Container();
+  }
+
+  @override
+  List<Widget> buildActions(BuildContext context) {
+    return <Widget>[
+      if (query.isNotEmpty)
+        IconButton(
+          tooltip: AppLocalizations.of(context)!.clearAll,
+          icon: const Icon(Icons.clear),
+          onPressed: () {
+            query = '';
+            showSuggestions(context);
+          },
+        ),
+    ];
+  }
+}
